+++ include/mousetrap/selection_model.hpp
@@ -67,7 +67,7 @@ namespace mousetrap
 
             /// @brief get indices of selected items
             /// @return vector of indices, may be empty
-            [[nodiscard]] std::vector<size_t> get_selection();
+            [[nodiscard]] std::vector<uint64_t> get_selection();
 
             /// @brief select all items or, if selection mode is single, select the first item
             void select_all();
@@ -78,19 +78,19 @@ namespace mousetrap
             /// @brief select item at specific index
             /// @param i index
             /// @param unselect_others if selection mode is multiple, should all other items be unselected
-            void select(size_t i, bool unselect_others = true);
+            void select(uint64_t i, bool unselect_others = true);
 
             /// @brief unselect item at specific item
             /// @param i index
-            void unselect(size_t i);
+            void unselect(uint64_t i);
 
             /// @brief if the selection mode is SelectionMode::SINGLE, determines whether no items can be selected
             /// @param b true to allow no selection
             void set_allow_no_selection_if_single(bool b);
 
             /// @brief get number of items
-            /// @return size_t
-            size_t get_n_items() const;
+            /// @return uint64_t
+            uint64_t get_n_items() const;
 
         protected:
             detail::SelectionModelInternal* _internal = nullptr;
